backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
web3>=7.0.0
eth-account>=0.11.0
siwe>=4.4.0
python-dotenv>=1.0.0
pydantic>=2.0.0
